[Add] Add Niagara Parameters Helper Function
#01. 나이아가라 유저 파라미터 적용 헬퍼 함수 추가
#02. 카미쉬 이펙트에 적용
