+++ Content/CR4S/_Blueprint/MonsterAI/BehaviorTree/Region/BT_KamishForestBoss.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:5057afba63b6302c3c5fcca3684c9ac402499cce7f79d3fd0e97a2f1eb22cd7d
-size 132145
+oid sha256:9176d03e54ba2445154ccc11a423631901803267d88e8c33c705a9d77634b389
+size 137701
