+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish1_Berserk_ThunderRush.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9f65455fc2fcccea8fbbfb12ea94e471e121db22a4e2c344d0b328fdaf8f8ae6
-size 47655
+oid sha256:51d02af82893ad9dfdac012a65e5dc43d47676a9f2ced896083750e8bb55a627
+size 47805
