+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish1_ThunderRush.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0a0160b3256f983d5c3b5305888bf42d5fb90139e37c0e2bf886f6b55e1cd7ea
-size 45726
+oid sha256:cf6b4a54573d83e867c8119f815d4150729625e86d41e0db216b3de91ef61ee4
+size 45830
