+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish2_Berserk_DashAndLeap.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9bd4c1d08eaf739e9ca8575a0cd2b0c7c0b1946199abd28cf5238eb7f5e44507
-size 35902
+oid sha256:6f16e87014dde00e3aaa183e511d06bbeba90d76b3b7dcbb46d29bc5dd470ea4
+size 35994
