+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish2_Berserk_KneelDash.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0757dca64330b9b93ccdcf4947b2c81aba6a0c5ad4cec5d565b97f1c9e15bc71
-size 32217
+oid sha256:236aa747d091872157fe76dd3376995237f13437be9209fdc665e8e75bf6a41b
+size 32186
