+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish2_Berserk_ThunderLeap.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:5c19f1e8f0659135d6896ee72dc3e839ff1e2729e7f7ef05af0ee6862a3079b1
-size 33829
+oid sha256:81525d9e69d52eca691fa19bc9cd09cd0a8c490d8657ae52a3e653daa5e855f0
+size 34146
