+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish2_DashAndLeap.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:a17404565d927944d825f45f6bcf6427e40b771ee92b0667f96f48a53214e21e
-size 28310
+oid sha256:22d125ba6104df834dc0321cf3d6c94b79bcb16103891fa45854b2e020b4f27d
+size 28391
