+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish2_KneelDash.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c1a754dab637deebdbb26a3d9a3686f25408144026855a78a8a335297282a768
-size 20403
+oid sha256:cabbdef6e6614fcb54cbb9826707471ce4324a18b24bd8dffa25fae02b1178fb
+size 20417
