+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish2_ThunderLeap.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:dac633774870bc75d4133a4c1a266e33b3e613796d0cdef8bd7204cb0697e0f6
-size 28053
+oid sha256:05297ac5f8ba39ce86a2c89ac0aa67b93b274c0a90858cc87ab921f928c04559
+size 28280
