+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish3_Berserk_FlyingHammer.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d63d69fcb91dc2891ea364b797bc28e95225e9bf273c41d5924c003b0f52ed3f
-size 21135
+oid sha256:c12bdd88ddfc7b0960973677d451cf3acc9c988cccbf56188e4dcfd652d6c1e6
+size 20685
