+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish3_FlyingHammer.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:83735f33d134be355c86915e402f3af014b5272cb92a578f409eac674e4c4b77
-size 21157
+oid sha256:469a1fa56b8128a12fe5583c7fc6e0124da55653885b467a7451f7b7b817d293
+size 21091
