+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish4_Berserk_RetrieveHammer.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:aea320271a08a84bef4f207bd9345b4531421a6a311ef6dafdde4fbcf84e01ad
-size 22393
+oid sha256:7a50013710156578ed6a92eb212d3b527821204eac6c55ed598008b75ebc59bd
+size 22539
