+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish4_RetrieveHammer.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b5ad06fe228a581e6bf92cf1fd593a7377caad03021f3c6428153633d6799344
-size 20672
+oid sha256:c476cbb7c8b998371853f917a2643e0fda2d31bb7c5156fb449c5e7b5d7099c0
+size 20845
