+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/Animations/AM_Kamish6_Berserk_HeavyStrike.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1522ed23c6152735f55ffbaf244b0801f33595fc5eed86603a59c4cd0ebaafc8
-size 18677
+oid sha256:bf4fc53e2af5de440b912564d1c1ab1e6c2f5cce8f806b1c697e56cd5ff1d939
+size 18632
