+++ Content/CR4S/_Blueprint/MonsterAI/BossMonster/Region/Kamish/BP_KamishForestBoss.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:eb3669b05c7fa7bd566acf35c4743b7fca6ec114eb39707d62941e2b26bceaab
-size 56822
+oid sha256:7ce78ba5ecf71443bd7db2308f48261b3623340d7692076848501c71c1fcfe4c
+size 56726
